refactor : 시작 클래스 명칭 변경에 따른 내용 수정
